- Code-cleanup for more proper compiling with less warnings.
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@3732 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/common/socket.c
@@ -90,7 +90,7 @@ static void setsocketopts(int fd)
 {
 	int yes = 1; // reuse fix
 	size_t buff;
-	unsigned int buff_size = sizeof (buff);	
+	size_t buff_size = sizeof (buff);	
 
 	setsockopt(fd,SOL_SOCKET,SO_REUSEADDR,(char *)&yes,sizeof yes);
 #ifdef SO_REUSEPORT
